fix: int round breaks coords on map

+++ components/map/map-results.tsx
@@ -61,8 +61,8 @@ function MapWithResults({
 
         new maplibregl.Marker({ color: constants.colors.frBlue })
           .setLngLat([
-            parseInt(result.siege.longitude),
-            parseInt(result.siege.latitude),
+            parseFloat(result.siege.longitude),
+            parseFloat(result.siege.latitude),
           ])
           .setPopup(popup)
           .addTo(map.current!);
@@ -86,7 +86,10 @@ function MapWithResults({
           new maplibregl.Marker({
             color: shouldColorZipCode ? 'yellow' : constants.colors.pastelBlue,
           })
-            .setLngLat([parseInt(match.longitude), parseInt(match.latitude)])
+            .setLngLat([
+              parseFloat(match.longitude),
+              parseFloat(match.latitude),
+            ])
             .setPopup(popup)
             .addTo(map.current!);
         }
